==> notebooks/3_preprocessing.ipynb <==
import pandas as pd
import numpy as np
import boto3

import sys

sys.path.append(
    "/Users/gabriel/Documents/Git/End-to-end-MLOps-for-Time-Series-Forecasting"
)
from utils import load_config

config = load_config("../config/development/config.yaml")

# Read data
data = pd.read_csv("../data/raw/train.csv")
reordered_columns = pd.Index(["target"]).append(data.columns.drop("target"))
data = data[reordered_columns]
print("data shape :", data.shape)
data.head(5)

len(data["data_block_id"].unique())

# Missing datetimes

# at each time step, a new value is generated per 'prediction_unit_id'
data = data.sort_values(by=["prediction_unit_id", "datetime"])
data.head()

data.insert(loc=6, column="datetime-1", value=data["datetime"].shift(1))
data.head()

data["datetime"] = pd.to_datetime(data["datetime"])
data["datetime-1"] = pd.to_datetime(data["datetime-1"])

data.insert(loc=7, column="timestep", value=data["datetime"] - data["datetime-1"])
data.head()

# identify discontinuous series, i.e., those which contain a timestep different from 1 hour
max_timestep_per_unit = (
    data[["prediction_unit_id", "timestep"]].groupby(by="prediction_unit_id").max()
)
discontinuous_series = max_timestep_per_unit[
    max_timestep_per_unit["timestep"] != pd.Timedelta(1, "h")
].index
max_timestep_per_unit.loc[discontinuous_series]

# drop discontinuous series
data = data[~data["prediction_unit_id"].isin(discontinuous_series)]

data["prediction_unit_id"].unique()

# Missing values

data.isna().sum()

data[data["target"].isna()].head()

mask = (
    (data["datetime"] <= "2021-10-31 04:00:00")
    & (data["datetime"] >= "2021-10-31 02:00:00")
    & (data["prediction_unit_id"] == 0)
)
data[mask]

We split consumption and production data to fill NaNs using the fill-forward method.

production = data.loc[
    data["is_consumption"] == 0, ["target", "datetime", "prediction_unit_id"]
]
consumption = data.loc[
    data["is_consumption"] == 1, ["target", "datetime", "prediction_unit_id"]
]

## Consumption

mask = (
    (consumption["datetime"] <= "2021-10-31 04:00:00")
    & (consumption["datetime"] >= "2021-10-31 02:00:00")
    & (consumption["prediction_unit_id"] == 0)
)
consumption[mask]

nan_consumption_indices = consumption[consumption["target"].isna()].index
nan_consumption_indices

consumption.isna().sum()

consumption.ffill(inplace=True)
consumption.isna().sum()

consumption[mask]

## Production

mask = (
    (production["datetime"] <= "2021-10-31 04:00:00")
    & (production["datetime"] >= "2021-10-31 02:00:00")
    & (production["prediction_unit_id"] == 0)
)
production[mask]

nan_production_indices = production[production["target"].isna()].index
nan_production_indices

production.isna().sum()

production.ffill(inplace=True)
production.isna().sum()

production[mask]

## Filling NaNs

nan_production_indices

nan_consumption_indices

nan_indices = nan_production_indices.join(nan_consumption_indices, how="outer")

data.loc[nan_indices, "target"].isna().sum() == data["target"].isna().sum()

data.loc[nan_consumption_indices, "target"] = consumption.loc[
    nan_consumption_indices, "target"
]
data.loc[nan_production_indices, "target"] = production.loc[
    nan_production_indices, "target"
]
data["target"].isna().sum()

# Downcasting

data.dtypes

int_columns = list(data.dtypes[data.dtypes == np.int64].index)
float_columns = list(data.dtypes[data.dtypes == np.float64].index)

int_columns

float_columns

for col in int_columns:
    data[col] = pd.to_numeric(data[col], downcast="unsigned")

for col in float_columns:
    data[col] = pd.to_numeric(data[col], downcast="float")

data.dtypes

# Train-Test split

data = data.sort_values(by="datetime")
data

data.iloc[int(len(data) * 0.80)]

train = data[data["datetime"] <= "2023-01-26 05:00:00"]
test = data[data["datetime"] > "2023-01-26 05:00:00"]
print(train.shape)
print(test.shape)

len(data) == len(train) + len(test)

len(train) / len(data)

len(test) / len(data)

# production_train = train[train["is_consumption"] == 0]
# production_test = test[test["is_consumption"] == 0]

# consumption_train = train[train["is_consumption"] == 1]
# consumption_test = test[test["is_consumption"] == 1]

# Save cleaned data

import os

if "preprocessed" not in os.listdir("../data"):
    os.mkdir("../data/preprocessed")

# TODO: save in an other format to preserve data types
train.to_csv("../data/preprocessed/train.csv", index=False)
# production_train.to_csv("../data/preprocessed/production_train.csv", index=False)
# consumption_train.to_csv("../data/preprocessed/consumption_train.csv", index=False)

test.to_csv("../data/preprocessed/test.csv", index=False)
# production_test.to_csv("../data/preprocessed/production_test.csv", index=False)
# consumption_test.to_csv("../data/preprocessed/consumption_test.csv", index=False)

os.listdir("../data/preprocessed")

# Load data to AWS S3 (Optional)
if config["s3_bucket"]:
    s3_client = boto3.client("s3")
    with open("../data/preprocessed/train.csv", "rb") as file:
        s3_client.upload_fileobj(
            file, config["s3_bucket"], "data/preprocessed/train.csv"
        )
    with open("../data/preprocessed/test.csv", "rb") as file:
        s3_client.upload_fileobj(
            file, config["s3_bucket"], "data/preprocessed/test.csv"
        )

# Check that data has been uploaded correctly
if config["s3_bucket"]:
    response = s3_client.list_objects(Bucket=config["s3_bucket"])
    for obj in response.get("Contents"):
        print(obj.get("Key"))
