==> notebooks/4_feature_engineering.ipynb <==
import pandas as pd

import sys

sys.path.append(
    "/Users/gabriel/Documents/Git/End-to-end-MLOps-for-Time-Series-Forecasting"
)
from utils import load_config

config = load_config("../config/config.yaml")

consumption_train = pd.read_csv(
    "../data/preprocessed/consumption_train.csv", parse_dates=["datetime"]
)
print(consumption_train.shape)
consumption_train.head()

# Lag features

def make_lags(df, column, lags, dropna=True):
    df_lags = pd.concat(
        {f"{column}_lag_{i}": df[column].shift(i) for i in lags}, axis=1
    )
    if dropna:
        return pd.concat([df, df_lags], axis=1).dropna(axis=0)
    else:
        return pd.concat([df, df_lags], axis=1)

start = config["forecast_horizon"]
end = start + config["n_lags"]
range(start, end)

groups = consumption_train.groupby(by="prediction_unit_id")
transformed_groups = []
for name, group in groups:
    group = make_lags(df=group, column="target", lags=range(start, end))
    transformed_groups.append(group)

new_consumption_train = pd.concat(transformed_groups, axis=0)
new_consumption_train.isna().sum().sum()

new_consumption_train.reset_index(drop=True, inplace=True)
new_consumption_train.index

new_consumption_train.head()

# Time-related features

def make_time_features(df, datetime_column):
    index = pd.DatetimeIndex(df[datetime_column])
    time_features = pd.DataFrame(
        index=index,
        data={"hour": index.hour, "dayofweek": index.dayofweek, "month": index.month},
    )
    time_features.reset_index(drop=True, inplace=True)
    return pd.concat([df, time_features], axis=1)

new_consumption_train = make_time_features(
    df=new_consumption_train, datetime_column="datetime"
)
print(new_consumption_train.shape)
new_consumption_train.head()

time_index = pd.DatetimeIndex(new_consumption_train["datetime"])
time_index

assert (time_index.hour == new_consumption_train["hour"]).sum() == len(
    new_consumption_train
)
assert (time_index.dayofweek == new_consumption_train["dayofweek"]).sum() == len(
    new_consumption_train
)
assert (time_index.month == new_consumption_train["month"]).sum() == len(
    new_consumption_train
)

new_consumption_train.drop(columns="datetime", inplace=True)

# Pipeline

def feature_engineering_pipeline(df):
    start = config["forecast_horizon"]
    end = start + config["n_lags"]
    groups = df.groupby(by="prediction_unit_id")
    transformed_groups = []
    for name, group in groups:
        group = make_lags(df=group, column="target", lags=range(start, end))
        transformed_groups.append(group)
    new_df = pd.concat(transformed_groups, axis=0)
    new_df.reset_index(drop=True, inplace=True)

    new_df = make_time_features(df=new_df, datetime_column="datetime")
    new_df.drop(columns="datetime", inplace=True)
    return new_df

df_ = feature_engineering_pipeline(consumption_train)

(df_ == new_consumption_train).sum().sum() == df_.size

del df_, new_consumption_train

# TODO: stack the last 'n_lags' rows from train on top of test so that we keep all rows from test after processing
consumption_test = pd.read_csv(
    "../data/preprocessed/consumption_test.csv", parse_dates=["datetime"]
)

consumption_train_processed = feature_engineering_pipeline(consumption_train)
consumption_test_processed = feature_engineering_pipeline(consumption_test)
print(consumption_train_processed.shape)
print(consumption_test_processed.shape)

# Save transformed data

import os

if "processed" not in os.listdir("../data"):
    os.mkdir("../data/processed")

# TODO: save in an other format to preserve data types
consumption_train_processed.to_csv(
    "../data/processed/consumption_train_processed.csv", index=False
)

consumption_test_processed.to_csv(
    "../data/processed/consumption_test_processed.csv", index=False
)

os.listdir("../data/processed")

# Load data to AWS S3 (Optional)
if config["s3_bucket"]:
    s3_client = boto3.client("s3")
    with open("../data/processed/train.csv", "rb") as file:
        s3_client.upload_fileobj(file, config["s3_bucket"], "data/processed/train.csv")
    with open("../data/processed/test.csv", "rb") as file:
        s3_client.upload_fileobj(file, config["s3_bucket"], "data/processed/test.csv")

# Check that data has been uploaded correctly
if config["s3_bucket"]:
    response = s3_client.list_objects(Bucket=config["s3_bucket"])
    for obj in response.get("Contents"):
        print(obj.get("Key"))
