==> notebooks/4_feature_engineering_mlforecast.ipynb <==
# Feature Engineering

In this notebook we will select and create features to feed our ML model. For now, we will focus on the consumption data.
<br>
We will use `MLForecast` to create the features.

import pandas as pd
from mlforecast import MLForecast
from mlforecast.lag_transforms import ExpandingMean, RollingMean
from pathlib import Path

import sys

sys.path.append(
    "/Users/gabriel/Documents/Git/End-to-end-MLOps-for-Time-Series-Forecasting"
)
from utils import load_config

default_config_path = "../configs/development/config.yaml"
config = load_config(default_config_path)

# Feature Engineering function


def feature_engineering(
    df,
    freq,
    id_col,
    time_col,
    target_col,
    lags,
    rolling_mean_window_sizes,
    lag_transforms_index,
    date_features,
    static_features,
    on_test=False,  # TODO: if True, stack last rows of train on top of test data to keep all test data
):
    lag_transforms = {
        i: [ExpandingMean()]
        + [RollingMean(window_size) for window_size in rolling_mean_window_sizes]
        for i in lag_transforms_index
    }
    fcst = MLForecast(
        models=[],
        freq=freq,
        lags=lags,
        lag_transforms=lag_transforms,
        date_features=date_features,
    )
    df_transformed = fcst.preprocess(
        df,
        id_col=id_col,
        time_col=time_col,
        target_col=target_col,
        static_features=static_features,
    )
    return df_transformed

# Testing the function

df = pd.read_csv("../data/preprocessed/consumption_train.csv", parse_dates=["datetime"])
print(df.shape)
df.head()

X = feature_engineering(
    df=df,
    freq=config["freq"],
    id_col=config["id_col"],
    time_col=config["time_col"],
    target_col=config["target_col"],
    lags=config["lags"],
    rolling_mean_window_sizes=config["rolling_mean_window_sizes"],
    lag_transforms_index=config["lag_transforms_index"],
    date_features=config["date_features"],
    static_features=config["static_features"],
    on_test=False,
)
X.head()

Explanation of the `lag_transforms` features:
- expanding_mean_lag**N**: expanding mean until lag **N**.
- rolling_mean_lag**N**_window_size**M**: rolling mean starting from lag **N** over the previous **M** values.
<br>
Example: rolling_mean_lag**5**_window_size**3** corresponds to the rolling mean starting from lag**5** over the previous **3** values. So this the means of lags 5, 6 and 7.

files = [
    "consumption_train.csv",
    "consumption_test.csv",
    "production_train.csv",
    "production_test.csv",
]

for file in files:
    on_test = True if "test" in file else False
    print(file, f": on_test={on_test}")

preprocessed_path = "../data/preprocessed"
processed_path = "../data/processed"

for file in files:
    df = pd.read_csv(Path(preprocessed_path, file), parse_dates=["datetime"])
    on_test = True if file.split(".")[0].split("_")[1] == "test" else False
    df_transformed = feature_engineering(
        df=df,
        freq=config["freq"],
        id_col=config["id_col"],
        time_col=config["time_col"],
        target_col=config["target_col"],
        lags=config["lags"],
        rolling_mean_window_sizes=config["rolling_mean_window_sizes"],
        lag_transforms_index=config["lag_transforms_index"],
        date_features=config["date_features"],
        static_features=config["static_features"],
        on_test=False,
    )
    # df_transformed.to_csv(Path(processed_path, file), index=False)
    print(f"{file} processed and saved to {processed_path}")

# Stack last train rows on top of test

When applying feature engineering we remove the first rows of the dataframe.
First determine the value of `step` for which `X` and `df` are aligned according to "datetime".
Once it's done, stack the last `n_step` values of `train` on top of `test` so that we don't lose
the first rows of test after preprocessing.

# # TODO: determine the value of `step` for which `X` and `df` are aligned according to "datetime".
# # Once it's done, stack the last `n_step` values of `train` on top of `test` so that we don't lose
# # the first rows of test after preprocessing.
# step = forecast_horizon + max(n_lags, n_lag_transforms) - 1
# step

# X_ = X.sort_values(by=["prediction_unit_id", "datetime"])
# df_ = df.sort_values(by=["prediction_unit_id", "datetime"])
# print(
#     (X_["datetime"].head().values == df_["datetime"].iloc[step : step + 5].values).sum()
#     == 5
# )
# print((X_["datetime"].tail().values == df_["datetime"].tail().values).sum() == 5)

# import random

# random.seed(0)

# forecast_horizon_list = random.choices(range(1, 10), k=3)
# n_lags_list = random.choices(range(1, 10), k=3)
# rolling_mean_window_size_list = random.choices(range(1, 10), k=3)
# n_lag_transforms_list = random.choices(range(1, 10), k=3)
# forecast_horizon_list

# import itertools

# combinations = itertools.product(
#     forecast_horizon_list,
#     n_lags_list,
#     rolling_mean_window_size_list,
#     n_lag_transforms_list,
# )
# # len(list(combinations))

# print(f"forecast_horizon_list : {forecast_horizon_list}")
# print(f"n_lags_list : {n_lags_list}")
# print(f"rolling_mean_window_size_list : {rolling_mean_window_size_list}")
# print(f"n_lag_transforms_list : {n_lag_transforms_list}")

# for _ in combinations:
#     forecast_horizon, n_lags, rolling_mean_window_size, n_lag_transforms = _
#     print(_)
#     break

# Testing the function on test set

# df_test = pd.read_csv("../data/preprocessed/consumption_test.csv", parse_dates=["datetime"])
# print(df_test.shape)
# df_test.head()

# X = feature_engineering(df_test, on_test=True)
# print(X.shape)
# X.head()

# X.tail()

# X.isna().sum().sum()
