==> notebooks/6_training.ipynb <==
from lightgbm import LGBMRegressor
from sklearn.metrics import mean_absolute_error
import mlflow
import time
from datetime import datetime, timezone

# import warnings
# warnings.filterwarnings("ignore", category=UserWarning)

import sys

sys.path.append(
    "/Users/gabriel/Documents/Git/End-to-end-MLOps-for-Time-Series-Forecasting"
)
from utils import load_config
from dotenv import dotenv_values

env_vars = dotenv_values("../.env")
config = load_config("../config/development/config.yaml")

# Read the Data

from utils import load_data

X_train, y_train = load_data("../data/processed/consumption_train.csv")
X_test, y_test = load_data("../data/processed/consumption_test.csv")

# Setting up the MLFlow experiment

From a terminal, run:
```
mlflow server --host 127.0.0.1 --port 5000
```

mlflow.set_tracking_uri("http://127.0.0.1:5000")

# Set the current active MLflow experiment
experiment_name = (
    f"Enefit DataV1 {config['model_name']} HpoConfigV{config['hpo_config_version']}"
)
try:
    mlflow.set_experiment(experiment_name=experiment_name)
except mlflow.exceptions.MlflowException:
    print("MLflow server not running. Please start the server using this command:")
    print("mlflow server --host 127.0.0.1 --port 5000")

experiment = mlflow.get_experiment_by_name(experiment_name)
print(f"ID for experiment '{experiment_name}': {experiment.experiment_id}")

# Log the study's best model

from utils import get_optuna_study

config["studies_dir"] = "../optuna_studies"
study = get_optuna_study(config)

# run metadata
utc_datetime = datetime.now(timezone.utc).strftime("%Y%m%dT%H%M%SUTC")
run_name = f"TrialNumber{study.best_trial.number}-{utc_datetime}"

# training
params = study.best_params
model = LGBMRegressor(**params, random_state=config["random_state"])
start = time.time()
model.fit(X_train, y_train)
end = time.time()

# metrics
y_fit = model.predict(X_train)
y_pred = model.predict(X_test)
train_mae = mean_absolute_error(y_train, y_fit)
test_mae = mean_absolute_error(y_test, y_pred)
training_duration = end - start
metrics = {
    "train_mae": train_mae,
    "test_mae": test_mae,
    "training_duration": training_duration,
}

with mlflow.start_run(run_name=run_name) as run:
    # Log the parameters used for the model fit
    mlflow.log_params(params)

    # Log the error metrics that were calculated during validation
    mlflow.log_metrics(metrics)

    # Log an instance of the trained model for later use
    mlflow.lightgbm.log_model(
        lgb_model=model,
        input_example=X_train.iloc[:1],
        artifact_path=config["model_name"],
    )
