==> notebooks/7_call_model.ipynb <==
import pandas as pd
import mlflow
import json
from preprocessing.preprocessing import *
preprocessing_version = 2

mlflow.__version__

model_name = "lightgbm"
preprocessing_version = 2

# Load and test the model

From a terminal, run:
```
mlflow server --host 127.0.0.1 --port 5000
```

mlflow.set_tracking_uri("http://127.0.0.1:5000")

df = pd.read_csv("./data/consumption.csv")
df["datetime"] = pd.to_datetime(df["datetime"])

print(f"Using preprocessing version {preprocessing_version}")
preprocessing = vars()[f"preprocessing_{preprocessing_version}"]
X, y = preprocessing(df)
print(f"X shape : {X.shape}")
print(f"y shape : {y.shape}")

## Validate the model before deployment

Run the following code to validate model inference works on the example input data and logged model dependencies, prior to deploying it to a serving endpoint

#BUG: NOT WORKING. Better to switch back to mlflow 2.17

# model_uri = 'runs:/963a54cea853490a9b21a465e6e99f94/lightgbm'
# # This is the input example logged with the model
# pyfunc_model = mlflow.pyfunc.load_model(model_uri)
# input_data = pyfunc_model.input_example

# # Verify the model with the provided input data using the logged dependencies.
# # For more details, refer to:
# # https://mlflow.org/docs/latest/models.html#validate-models-before-deployment
# mlflow.models.predict(
#     model_uri=model_uri,
#     input_data=input_data,
#     # env_manager="uv",
# )

logged_model = 'runs:/963a54cea853490a9b21a465e6e99f94/lightgbm'

# Load model as a PyFuncModel.
loaded_model = mlflow.pyfunc.load_model(logged_model)
loaded_model.predict(X[:1])

# Deploy the model

from mlflow.tracking import MlflowClient

client = MlflowClient()
experiment_name = f"Enefit {model_name} Preprocessing {preprocessing_version}"
experiment = client.get_experiment_by_name(experiment_name)
experiment_id = experiment.experiment_id
print(f"Run ID : {experiment_id}")

# Fetch the most recent run
runs = client.search_runs([experiment_id], order_by=["start_time DESC"], max_results=10)
# runs[0].to_dictionary()["data"]["metrics"]#["test_mae"]
if runs:
    latest_run_id = runs[0].info.run_id
    print(f"Latest Run ID: {latest_run_id}")

PORT = 1234
print(
    f"""Run the below command in a new window. You must be in the same repo as your mlruns directory and have mlflow installed :

    mlflow models serve -m ./mlartifacts/{experiment_id}/{latest_run_id}/artifacts/lightgbm -p {PORT}"""
)

# Call the deployed model

import requests
import json

address = "127.0.0.1"
# address = "3.252.192.81"
# address = "ec2-54-228-144-127.eu-west-1.compute.amazonaws.com"


# Check if the model responds
url = f"http://{address}:{PORT}/ping"
r = requests.get(url)
print(r)
print(r.text)

# See https://mlflow.org/docs/latest/deployment/deploy-model-locally.html#json-input
json_data = json.loads(X.iloc[:1, :5].to_json(orient="split"))
json_data.pop('index', None)
json_data = {"dataframe_split": json_data}
json_data = json.dumps(json_data, indent=4)
print(json_data)

json_data = json.loads(X.iloc[:1, :].to_json(orient="split"))
json_data.pop('index', None)
json_data = {"dataframe_split": json_data}
json.dumps(json_data)

# get predictions
url = f"http://{address}:{PORT}/invocations"
r = requests.post(url, headers={"Content-Type": "application/json"}, json=json_data)
r
