==> notebooks/7_save_best_run.ipynb <==
import mlflow
from mlflow import MlflowClient
from dotenv import dotenv_values

import sys

env_vars = dotenv_values("../.env")
sys.path.append(f"{env_vars['WORKING_DIR']}")
from utils import load_config, save_config

config = load_config("../config/development/config.yaml")

# Retrieve runs

From a terminal, in the same repo as your mlruns directory, run:
```
mlflow server --host 127.0.0.1 --port 5000
```

client = MlflowClient(
    tracking_uri=env_vars["MLFLOW_TRACKING_URI"],
)

experiment_name = f"Enefit DataV1 lightgbm HpoConfigV1"
experiment = client.get_experiment_by_name(experiment_name)
experiment_id = experiment.experiment_id
runs_df = mlflow.search_runs([experiment_id])
runs_df.head()

sorted_runs_df = runs_df.sort_values(by="metrics.test_mae", ascending=True)
sorted_runs_df.head()

best_run = sorted_runs_df.iloc[0]
best_run_id = best_run["run_id"]
best_run_id

import mlflow

logged_model = f"runs:/{best_run_id}/lightgbm"

# Load model as a PyFuncModel.
loaded_model = mlflow.pyfunc.load_model(logged_model)

# # Predict on a Pandas DataFrame.
# import pandas as pd

# loaded_model.predict(pd.DataFrame(data))

loaded_model

loaded_model.get_raw_model()

# Save the model

import joblib
from utils import create_dir

# Create the models directory if it doesn't exist
create_dir("models")

# Save the model to a file
with open(f"models/{config['model_name']}.joblib", "wb") as f:
    joblib.dump(loaded_model, f)

# Load the model from the file
with open(f"models/{config['model_name']}.joblib", "rb") as f:
    loaded_model = joblib.load(f)

loaded_model

# Load metadata

import json

metadata = best_run.to_json(indent=4)
with open(f"models/{config['model_name']}_metadata.json", "w") as f:
    f.write(metadata)

with open(f"models/{config['model_name']}_metadata.json", "r") as f:
    metadata = json.load(f)
metadata

config["EXPERIMENT_ID"] = experiment_id
config["RUN_ID"] = best_run_id

config

save_config(config, "../config/development/config.yaml")

loaded_config = load_config("../config/development/config.yaml")
config == loaded_config
