==> notebooks/8_call_model.ipynb <==
import pandas as pd
import mlflow
from dotenv import dotenv_values

import sys

env_vars = dotenv_values("../.env")
sys.path.append(f"{env_vars['WORKING_DIR']}")
from utils import load_config

config = load_config("../config/development/config.yaml")

# Load the Data

from lgbm_hpo import load_data

X_train, y_train = load_data("../data/processed/consumption_train.csv")
X_test, y_test = load_data("../data/processed/consumption_test.csv")

# Load and test the model

From a terminal, run:
```
mlflow server --host 127.0.0.1 --port 5000
```

mlflow.set_tracking_uri("http://127.0.0.1:5000")

## Validate the model before deployment

Run the following code to validate model inference works on the example input data and logged model dependencies, prior to deploying it to a serving endpoint

EXPERIMENT_ID = config["EXPERIMENT_ID"]
RUN_ID = config["RUN_ID"]

logged_model = f"runs:/{RUN_ID}/lightgbm"

# Load model as a PyFuncModel.
loaded_model = mlflow.pyfunc.load_model(logged_model)
loaded_model.predict(X_train[:1])

# Deploy the model with MLflow

PORT = 5050
print(
    f"""Run the below command in a new window. You must be in the same repo as your mlruns directory and have mlflow installed :

    mlflow models serve -m ./mlartifacts/{EXPERIMENT_ID}/{RUN_ID}/artifacts/lightgbm -p {PORT}"""
)

import requests
import json

address = "127.0.0.1"
PORT = 5001
# address = "3.252.192.81"
# address = "ec2-54-228-144-127.eu-west-1.compute.amazonaws.com"

# Check if the model responds
url = f"http://{address}:{PORT}/ping"
r = requests.get(url)
print(r)
print(r.text)

# See https://mlflow.org/docs/latest/deployment/deploy-model-locally.html#json-input
json_data = json.loads(X_train.iloc[:1, :5].to_json(orient="split"))
json_data.pop("index", None)
json_data = {"dataframe_split": json_data}
json_data = json.dumps(json_data, indent=4)
print(json_data)

# See https://mlflow.org/docs/latest/deployment/deploy-model-locally.html#json-input
json_data = json.loads(X_train.iloc[:1, :].to_json(orient="split"))
json_data.pop("index", None)
json_data = {"dataframe_split": json_data}
# json_data = json.dumps(json_data, indent=4)
# print(json_data)
json_data

# get predictions
url = f"http://{address}:{PORT}/invocations"
r = requests.post(url, headers={"Content-Type": "application/json"}, json=json_data)
r

type(json_data)

# Deploy the model with Flask

Run the `deployment.py` script in the repo's root directory

import requests
import json

address = "127.0.0.1"
PORT = 8080
# address = "3.252.192.81"
# address = "ec2-54-228-144-127.eu-west-1.compute.amazonaws.com"

# Check if the model responds
url = f"http://{address}:{PORT}/"
r = requests.get(url)
print(r)
print(r.text)

model_uri = f"runs:/{RUN_ID}/lightgbm"
# The model is logged with an input example
pyfunc_model = mlflow.pyfunc.load_model(model_uri)
input_data = pyfunc_model.input_example

json_data = json.loads(input_data.to_json(orient="split"))
json_data.pop("index", None)
json_data = {"dataframe_split": json_data}
# json_data = json.dumps(json_data, indent=4)
# print(json_data)
# json_data

import joblib

with open(f"models/{config['model_name']}.joblib", "rb") as f:
    model = joblib.load(f)

json_data["dataframe_split"].keys()

pd.DataFrame(
    data=json_data["dataframe_split"]["data"],
    columns=json_data["dataframe_split"]["columns"],
)

# get predictions
url = f"http://{address}:{PORT}/predict"
r = requests.post(url, headers={"Content-Type": "application/json"}, json=json_data)
r

print(r.json())
